add smsSend class
